summer day3
ADDED:
Created an insert function in collection that inserts the item that has been clicked to the database in the pet table.
-Result: It now remembers what the pet was equiping even if you reopen the game after closing it. because it now just retrieves the item data equiped from the pet table.

I added more 2 duplicates of the clickable_item.scene and script for both "collection" and "shop".

Added functionality that the shop now removes the item that are owned by the player in the item display.

FIXED:
Fixed the change scene button where it still detects input even if it is no visible.

CHANGED:
Changed the term "part" to "item".

Change "_fashion()" function

REMOVED:
Term "part removed.

Removed  pet_customize scene

Removed the old way of changing pet texture functionality.